pyproject.toml
[tool.poetry]
name = "standort-app"
version = "0.1.1"
description = "A portable API for Airtables"
authors = ["Oleg Lavrovsky <oleg@datalets.ch>"]

[tool.poetry.dependencies]
python = ">=3.6,<4"
gunicorn = "*"
python-dotenv = "*"
flask_api = "^2.0"
Airtable = "^0.3.1"
urllib3 = ">=1.24.2"
